feat: Add required software information to installation instructions
This commit adds information about the required software for installing and using the template. It includes sections for:

-   Typst (official code editor)
-   Visual Studio Code (code editor)
-   Tinymist Typst Visual Studio Code Extension

This improves the clarity of the installation instructions by explicitly stating the necessary software and tools. The changes include translations for both Turkish and English. Additionally, the commit updates links to the Visual Studio Code and Tinymist Typst extension.