feat: Add import all variables comment to README
This commit adds a comment to the README.md file that explains how to import all variables from the template using `#import ...: *` instead of specifying them individually.

The comment is added to the code examples for different usage scenarios:
- Using the template from the online preview.
- Using the template by downloading it to the location mentioned in the documentation.
- Using the template by manually downloading it to any location.

This enhancement aims to improve the template's usability by providing users with a more concise way to import variables.