feat: Add "Getting Started" sections to README
This commit adds "Getting Started" sections to the README.md file in both Turkish and English. These sections provide a brief introduction on how to use the template after importing it, emphasizing the need to configure the template settings. It also clarifies that when the template is used as described in the Installation section, this part is automatically available, and users only need to update the template configurations. This enhancement aims to guide new users and improve the onboarding experience.