refactor: Move "Selecting one of the alternatives in headings with alternatives" section
This commit moves the "Selecting one of the alternatives in headings with alternatives" section from the "Features" section to the "Usage" section in the README.md file. This change improves the organization of the document by placing the usage instructions closer to the "Getting Started" section, making it easier for users to find relevant information.