fix: Correct grammar in README collapsible summaries
This commit fixes grammatical errors in the summaries of the collapsible sections within the README.md file. Specifically, it changes "Click to show/hide code" and "Click to show/hide diagram" to "Click to show/hide the code" and "Click to show/hide the diagram" respectively. This improves the clarity and correctness of the instructions for users interacting with the collapsible sections.