fix: Correct pluralization in README regarding official templates
This commit fixes a pluralization error in the README.md file. The word "şablonu" (template) was incorrectly used in the singular form when referring to multiple templates. This has been corrected to "şablonları" (templates) to accurately reflect that multiple templates are available in the specified format. This change improves the grammatical accuracy and clarity of the README.