fix: Correct URLs for doctorate and non-thesis master's templates in README
This commit fixes incorrect URLs in the README.md file for the doctorate (PhD) and evening education non-thesis master's degree templates. The previous URLs pointed to the thesis master's degree templates page. The URLs have been updated to link to the correct pages for the respective programs, ensuring users are directed to the appropriate resources.