fix: Update author field comments in template files
This commit updates the comments associated with the author field in both `src/template-configurations.typ` and `template/main.typ`. The comment is changed from "Öğrencinin adı, soyadı ve ORCID numarası. [The student's name, lastname and ORCID number.]" to "Yazarın (Öğrencinin) adı, soyadı ve ORCID numarası. [The author's (student's) name, lastname and ORCID number.]". This change clarifies that the field refers to the author, which is typically the student, and improves the overall clarity and accuracy of the template configurations.