feat: Add parameter usage explanations to README
This commit adds explanations on how to set the values of the parameters in the template configurations. The explanations are added as comment lines above the relevant parameter in both the Turkish and English sections of the README. This provides clearer guidance to users on how to configure the template.