feat: Add "beni-oku.md" files to font directories
This commit adds "beni-oku.md" files to the following font directories within the template's assets:

- template/assets/fonts/tex-gyre-termes-math/
- template/assets/fonts/tex-gyre-termes/
- template/assets/fonts/cambria-math/
- template/assets/fonts/times-new-roman/

These files provide instructions to users on how to add the corresponding static font files to the respective directories. They also include licensing information for open-source fonts (TeX Gyre Termes and TeX Gyre Termes Math) and a note indicating that this step is optional, with a link to the package information on typst.app.