refactor: Rename equation list to math equation list
This commit renames all instances of "equation" to "math equation" to provide a more accurate and descriptive name for the list of mathematical equations. This change affects the following files:

- `src/template-configurations.typ`
- `src/sections/thesis-front-section.typ`
- `template/main.typ`
- `src/core/validation-manager/validation-manager.typ`
- `src/sections/01-front/list-of-equations-page.typ` (renamed to `list-of-math-equations-page.typ`)
- `src/styles/list-of-equations-style.typ` (renamed to `list-of-math-equations-style.typ`)

The purpose of this change is to improve the clarity and maintainability of the codebase by using more precise terminology.