feat: Add font information to README
This commit adds a new section to the README file that provides information about the fonts used in the template. It explains the licensing restrictions of 'Times New Roman' and 'Cambria Math', and how 'TeX Gyre Termes' and 'TeX Gyre Termes Math' are used as alternatives. It also provides instructions on how to use the original fonts if the user has the legal right to do so, and how to install them for Tinymist Typst users.