refactor: Reorder imports for consistency across multiple files
This commit reorders the import statements in several `.typ` files to improve consistency and readability. The primary goal is to group related imports together, making it easier to understand the dependencies of each file.

The following files were modified:

- `src/core/validation-manager/validators/keywords-validator.typ`
- `src/sections/03-back/curriculum-vitae-page.typ`
- `src/sections/01-front/thesis-study-originality-report-page.typ`
- `src/core/validation-manager/validators/work-packages-validator.typ`
- `src/styles/thesis-main-section-heading-style.typ`
- `src/sections/01-front/scientific-ethics-declaration-page.typ`
- `src/sections/01-front/preface-page.typ`
- `src/components/full-date-with-author-fullname-component.typ`
- `src/sections/03-back/work-schedule-page.typ`
- `src/styles/list-of-math-equations-style.typ`
- `src/styles/thesis-back-section-heading-style.typ`
- `src/styles/thesis-front-section-heading-style.typ`
- `src/styles/list-of-image-figures-style.typ`
- `src/styles/list-of-code-figures-style.typ`
- `src/styles/list-of-table-figures-style.typ`

In each of these files, the import statements were reordered to group related imports together, such as language-related imports or imports from the same directory. This change does not affect the functionality of the code, but it improves its overall organization and maintainability.